+++ fixtures/c/linked_list.c
@@ -1,7 +1,10 @@
 #include <stdlib.h>
 
 int main() {
-    int *head; int *temp; int *node;
+    // M31: cells must be `long` (8 bytes) so a single slot can hold
+    // either a value or a pointer-shaped next link. With `int` cells
+    // the next-pointer cast would truncate to 32 bits.
+    long *head; long *temp; long *node;
     int sum; int i;
 
     head = 0;
@@ -10,12 +13,12 @@ int main() {
     // Create a list of 5 nodes: [4, 3, 2, 1, 0]
     for (i = 0; i < 5; i++) {
         // One slot for value, one for the next pointer. Stored as a
-        // flat int[2]; the cast on the next-pointer write tells the
+        // flat long[2]; the cast on the next-pointer write tells the
         // type-checker we know we're stuffing a pointer-shaped value
         // into a word-sized cell.
-        node = malloc(sizeof(int) + sizeof(int *));
+        node = malloc(sizeof(long) + sizeof(long *));
         node[0] = i;            // data
-        node[1] = (int)head;    // next
+        node[1] = (long)head;   // next
         head = node;
     }
 
@@ -23,7 +26,7 @@ int main() {
     temp = head;
     while (temp != 0) {
         sum = sum + temp[0];
-        temp = (int *)temp[1];
+        temp = (long *)temp[1];
     }
 
     return sum; // Expected: 4+3+2+1+0 = 10
